feat: move functions into compsoables
