+++ apps/web/composables/useProduct/useProduct.ts
@@ -1,4 +1,5 @@
 import type { Product, ProductParams } from '@plentymarkets/shop-api';
+import { categoryTreeGetters, productGetters } from '@plentymarkets/shop-sdk';
 import { toRefs } from '@vueuse/shared';
 import type { UseProductReturn, UseProductState, FetchProduct } from '~/composables/useProduct/types';
 import { useSdk } from '~/sdk';
@@ -16,6 +17,7 @@ export const useProduct: UseProductReturn = (slug) => {
   const state = useState<UseProductState>(`useProduct-${slug}`, () => ({
     data: {} as Product,
     loading: false,
+    breadcrumbs: []
   }));
 
   /** Function for fetching product data.
@@ -39,7 +41,38 @@ export const useProduct: UseProductReturn = (slug) => {
     return state.value.data;
   };
 
+  /**
+   * @description Function for generating breadcrumbs
+   * @example generateBreadcrumbs()
+   */
+  const generateBreadcrumbs = () => {
+    const { data: categoryTree } = useCategoryTree();
+    const { t } = useI18n();
+    const breadcrumb = categoryTreeGetters.generateBreadcrumbFromCategory(
+      categoryTree.value,
+      Number(productGetters.getCategoryIds(state.value.data)?.[0] ?? 0),
+    );
+    breadcrumb.unshift({ name: t('home'), link: '/' });
+    breadcrumb.push({ name: productGetters.getName(state.value.data), link: `#` });
+    state.value.breadcrumbs = breadcrumb;
+  };
+
+  /**
+   * @description Function for setting product title meta data
+   */
+  const setTitle = () => {
+    const title = productGetters.getName(state.value.data);
+
+    useHead({
+      title
+    });
+  };
+
+  
+
   return {
+    setTitle,
+    generateBreadcrumbs,
     fetchProduct,
     ...toRefs(state.value),
   };
